Removing set and get schema, moving into single subcommand
